[REFACTOR] Split up new shading code
